+++ Model/SessionManager.php
@@ -3,7 +3,7 @@
 namespace FlowCommerce\FlowConnector\Model;
 
 use FlowCommerce\FlowConnector\Api\SessionManagementInterface;
-use Magento\Checkout\Model\Session;
+use Magento\Checkout\Model\Session as CheckoutSession;
 use Magento\Customer\Api\AddressRepositoryInterface;
 use Magento\Customer\Model\Session as CustomerSession;
 use Magento\Framework\Stdlib\Cookie\CookieMetadataFactory;
@@ -11,6 +11,8 @@ use Magento\Framework\Session\SessionManagerInterface;
 use Magento\Framework\Stdlib\CookieManagerInterface;
 use FlowCommerce\FlowConnector\Model\Api\Session as ApiSession;
 use Magento\Quote\Model\Quote;
+use FlowCommerce\FlowConnector\Model\Api\Order\Get as OrderGet;
+use Psr\Log\LoggerInterface;
 
 class SessionManager implements SessionManagementInterface
 {
@@ -23,6 +25,11 @@ class SessionManager implements SessionManagementInterface
      * Name of Flow session cookie
      */
     const FLOW_SESSION_COOKIE = '_f60_session';
+
+    /**
+     * Name of Flow session cookie update flag
+     */
+    const FLOW_SESSION_UPDATE_COOKIE_FLAG = 'flow_mage_session_update';
 
     /**
      * @var SessionManagerInterface
@@ -45,11 +52,6 @@ class SessionManager implements SessionManagementInterface
     private $sessionApiClient;
 
     /**
-     * @var Session
-     */
-    private $session;
-
-    /**
      * @var Configuration
      */
     private $configuration;
@@ -60,7 +62,7 @@ class SessionManager implements SessionManagementInterface
     private $customerSession;
 
     /**
-     * @var Session
+     * @var CheckoutSession
      */
     private $checkoutSession;
 
@@ -70,55 +72,75 @@ class SessionManager implements SessionManagementInterface
     private $addressRepository;
 
     /**
+     * @var FlowCartManager
+     */
+    private $flowCartManager;
+
+    /*
+     * @var OrderGet
+     */
+    private $orderGet;
+
+    /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
+    /**
      * SessionManager constructor.
      * @param CookieManagerInterface $cookieManagerInterface
      * @param CookieMetadataFactory $cookieMetadataFactory
      * @param SessionManagerInterface $sessionManagerInterface
      * @param ApiSession $sessionApiClient
-     * @param Session $session
      * @param \FlowCommerce\FlowConnector\Model\Configuration $configuration
      * @param CustomerSession $customerSession
-     * @param Session $checkoutSession
+     * @param CheckoutSession $checkoutSession
      * @param AddressRepositoryInterface $addressRepository
+     * @param OrderGet $orderGet
+     * @param LoggerInterface $logger
      */
     public function __construct(
         CookieManagerInterface $cookieManagerInterface,
         CookieMetadataFactory $cookieMetadataFactory,
         SessionManagerInterface $sessionManagerInterface,
         ApiSession $sessionApiClient,
-        Session $session,
         Configuration $configuration,
         CustomerSession $customerSession,
-        Session $checkoutSession,
-        AddressRepositoryInterface $addressRepository
+        CheckoutSession $checkoutSession,
+        AddressRepositoryInterface $addressRepository,
+        OrderGet $orderGet,
+        LoggerInterface $logger
     ) {
 
         $this->cookieManagerInterface = $cookieManagerInterface;
         $this->cookieMetadataFactory = $cookieMetadataFactory;
         $this->sessionManagerInterface = $sessionManagerInterface;
         $this->sessionApiClient = $sessionApiClient;
-        $this->session = $session;
         $this->configuration = $configuration;
         $this->customerSession = $customerSession;
         $this->checkoutSession = $checkoutSession;
         $this->addressRepository = $addressRepository;
+        $this->orderGet = $orderGet;
+        $this->logger = $logger;
     }
 
     /**
      * Retrieve data for Flow session in progress
-     *
      * @return array|null
+     * @throws \Magento\Framework\Exception\InputException
      * @throws \Magento\Framework\Exception\NoSuchEntityException
+     * @throws \Magento\Framework\Stdlib\Cookie\FailureToSendException
      */
     public function getFlowSessionData()
     {
-        $flowSessionData = $this->session->getFlowSessionData();
-        if (!$flowSessionData) {
+        $flowSessionData = $this->checkoutSession->getFlowSessionData();
+        if (!$flowSessionData || $this->isUpdateSession()) {
             $sessionId = $this->cookieManagerInterface->getCookie(self::FLOW_SESSION_COOKIE);
             if ($sessionId) {
                 $flowSessionData = $this->sessionApiClient->getFlowSessionData($sessionId);
                 if ($flowSessionData) {
-                    $this->session->setFlowSessionData($flowSessionData);
+                    $this->setFlowSessionData($flowSessionData);
+                    $this->cookieManagerInterface->deleteCookie(self::FLOW_SESSION_UPDATE_COOKIE_FLAG);
                 }
             }
         }
@@ -152,7 +174,7 @@ class SessionManager implements SessionManagementInterface
                 $cookieMeta
             );
 
-            $this->session->setFlowSessionData($flowSessionData);
+            $this->checkoutSession->setFlowSessionData($flowSessionData);
         }
     }
 
@@ -168,12 +190,11 @@ class SessionManager implements SessionManagementInterface
     {
         $quote = $this->checkoutSession->getQuote();
 
-        $items = $quote->getItems();
-        if (!$items) {
+        $items = $quote->getAllVisibleItems();
+        if (!$items || !$country) {
             return null;
         }
 
-        $url = $this->configuration->getFlowCheckoutUrl() . '?';
 
         // Additional custom attributes to pass through hosted checkout
         $attribs = [];
@@ -186,6 +207,12 @@ class SessionManager implements SessionManagementInterface
             $params['country'] = $country;
         }
 
+        // Removed, unnecessary for now
+        /* $flowOrderNumber = $quote->getFlowConnectorOrderNumber(); */
+        /* $query = ['number' => $flowOrderNumber]; */
+        /* $result = $this->orderGet->execute($query); */
+        /* $flowOrderData = reset($result); */
+        
         if ($flowSessionId = $this->cookieManagerInterface->getCookie(self::FLOW_SESSION_COOKIE)) {
             $params['flow_session_id'] = $flowSessionId;
         }
@@ -229,7 +256,19 @@ class SessionManager implements SessionManagementInterface
         $ctr = 0;
         foreach ($items as $item) {
             $params['items[' . $ctr . '][number]'] = $item->getSku();
+            $itemRowTotal = $item->getRowTotal();
+            $itemDiscountAmount = $item->getDiscountAmount();
+            $itemDiscountPercentage = 0.0;
+            if ($itemRowTotal > 0 && $itemDiscountAmount > 0) {
+                $itemDiscountPercentage = (float)(($itemDiscountAmount / $itemRowTotal) * 100);
+            }
+
             $params['items[' . $ctr . '][quantity]'] = $item->getQty();
+            if ($itemDiscountPercentage > 0) {
+                $params['items[' . $ctr . '][discounts][discounts][0][offer][discriminator]'] = 'discount_offer_percent';
+                $params['items[' . $ctr . '][discounts][discounts][0][offer][percent]'] = $itemDiscountPercentage;
+                $params['items[' . $ctr . '][discounts][discounts][0][label]'] = 'Discount';
+            }
             $ctr += 1;
         }
 
@@ -238,19 +277,55 @@ class SessionManager implements SessionManagementInterface
             $params['attributes[' . $k . ']'] = $v;
         }
 
-        $url = $url . http_build_query($params);
-        return $url;
+        $this->logger->info('CART: ' . json_encode($params));
+        return $this->configuration->getFlowCheckoutUrl() . '?' . http_build_query($params);
     }
 
     /**
      * Get session experience country
-     * @return string|null $country
+     * @return string|null
+     * @throws \Magento\Framework\Exception\InputException
      * @throws \Magento\Framework\Exception\NoSuchEntityException
+     * @throws \Magento\Framework\Stdlib\Cookie\FailureToSendException
      */
     public function getSessionExperienceCountry()
     {
         $sessionData = $this->getFlowSessionData();
         $country = isset($sessionData['experience']['country']) ? $sessionData['experience']['country'] : null;
         return $country;
+    }
+
+    /**
+     * Get session experience country
+     * @return string|null
+     * @throws \Magento\Framework\Exception\InputException
+     * @throws \Magento\Framework\Exception\NoSuchEntityException
+     * @throws \Magento\Framework\Stdlib\Cookie\FailureToSendException
+     */
+    public function getSessionExperienceKey()
+    {
+        $sessionData = $this->getFlowSessionData();
+        if ($country = isset($sessionData['experience']['key']) ? $sessionData['experience']['key'] : null) {
+            $this->logger->info('Current Experience Key: '.$country);
+        }
+        return $country;
+    }
+
+    /**
+     * Set flow session data
+     * @param $data
+     */
+    public function setFlowSessionData($data)
+    {
+        $this->checkoutSession->setFlowSessionData($data);
+    }
+
+    /**
+     * Return cookie flag if session is to be updated
+     * @return bool
+     */
+    private function isUpdateSession()
+    {
+        return (bool)$this->cookieManagerInterface->getCookie(self::FLOW_SESSION_UPDATE_COOKIE_FLAG);
     }
 }